Migrate to null-object pattern.
