+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/internal/state/DefaultSynchronousMetricStorage.java
@@ -1,0 +1,153 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.sdk.metrics.internal.state;
+
+import io.opentelemetry.api.common.Attributes;
+import io.opentelemetry.context.Context;
+import io.opentelemetry.sdk.metrics.data.MetricData;
+import io.opentelemetry.sdk.metrics.internal.aggregator.Aggregator;
+import io.opentelemetry.sdk.metrics.internal.aggregator.AggregatorHandle;
+import io.opentelemetry.sdk.metrics.internal.descriptor.MetricDescriptor;
+import io.opentelemetry.sdk.metrics.internal.view.AttributesProcessor;
+import java.util.Map;
+import java.util.Objects;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.locks.ReentrantLock;
+
+/**
+ * Stores aggregated {@link MetricData} for synchronous instruments.
+ *
+ * <p>This class is internal and is hence not for public use. Its APIs are unstable and can change
+ * at any time.
+ */
+public final class DefaultSynchronousMetricStorage<T> implements SynchronousMetricStorage {
+  private final MetricDescriptor metricDescriptor;
+  private final ConcurrentHashMap<Attributes, AggregatorHandle<T>> aggregatorLabels;
+  private final ReentrantLock collectLock;
+  private final Aggregator<T> aggregator;
+  private final InstrumentProcessor<T> instrumentProcessor;
+  private final AttributesProcessor attributesProcessor;
+
+  DefaultSynchronousMetricStorage(
+      MetricDescriptor metricDescriptor,
+      Aggregator<T> aggregator,
+      InstrumentProcessor<T> instrumentProcessor,
+      AttributesProcessor attributesProcessor) {
+    this.metricDescriptor = metricDescriptor;
+    aggregatorLabels = new ConcurrentHashMap<>();
+    collectLock = new ReentrantLock();
+    this.aggregator = aggregator;
+    this.instrumentProcessor = instrumentProcessor;
+    this.attributesProcessor = attributesProcessor;
+  }
+
+  // This is a storage handle to use when the attributes processor requires
+  private final BoundStorageHandle lateBoundStorageHandle =
+      new BoundStorageHandle() {
+        @Override
+        public void release() {}
+
+        @Override
+        public void recordLong(long value, Attributes attributes, Context context) {
+          DefaultSynchronousMetricStorage.this.recordLong(value, attributes, context);
+        }
+
+        @Override
+        public void recordDouble(double value, Attributes attributes, Context context) {
+          DefaultSynchronousMetricStorage.this.recordDouble(value, attributes, context);
+        }
+      };
+
+  @Override
+  public BoundStorageHandle bind(Attributes attributes) {
+    Objects.requireNonNull(attributes, "attributes");
+    if (attributesProcessor.usesContext()) {
+      // We cannot pre-bind attributes because we need to pull attributes from context.
+      return lateBoundStorageHandle;
+    }
+    return doBind(attributesProcessor.process(attributes, Context.current()));
+  }
+
+  private BoundStorageHandle doBind(Attributes attributes) {
+    AggregatorHandle<T> aggregatorHandle = aggregatorLabels.get(attributes);
+    if (aggregatorHandle != null && aggregatorHandle.acquire()) {
+      // At this moment it is guaranteed that the Bound is in the map and will not be removed.
+      return aggregatorHandle;
+    }
+
+    // Missing entry or no longer mapped, try to add a new entry.
+    aggregatorHandle = aggregator.createHandle();
+    while (true) {
+      AggregatorHandle<?> boundAggregatorHandle =
+          aggregatorLabels.putIfAbsent(attributes, aggregatorHandle);
+      if (boundAggregatorHandle != null) {
+        if (boundAggregatorHandle.acquire()) {
+          // At this moment it is guaranteed that the Bound is in the map and will not be removed.
+          return boundAggregatorHandle;
+        }
+        // Try to remove the boundAggregator. This will race with the collect method, but only one
+        // will succeed.
+        aggregatorLabels.remove(attributes, boundAggregatorHandle);
+        continue;
+      }
+      return aggregatorHandle;
+    }
+  }
+
+  // Overridden to make sure attributes processor can pull baggage.
+  @Override
+  public void recordLong(long value, Attributes attributes, Context context) {
+    Objects.requireNonNull(attributes, "attributes");
+    attributes = attributesProcessor.process(attributes, context);
+    BoundStorageHandle handle = doBind(attributes);
+    try {
+      handle.recordLong(value, attributes, context);
+    } finally {
+      handle.release();
+    }
+  }
+
+  // Overridden to make sure attributes processor can pull baggage.
+  @Override
+  public void recordDouble(double value, Attributes attributes, Context context) {
+    Objects.requireNonNull(attributes, "attributes");
+    attributes = attributesProcessor.process(attributes, context);
+    BoundStorageHandle handle = doBind(attributes);
+    try {
+      handle.recordDouble(value, attributes, context);
+    } finally {
+      handle.release();
+    }
+  }
+
+  @Override
+  public MetricData collectAndReset(long startEpochNanos, long epochNanos) {
+    collectLock.lock();
+    try {
+      for (Map.Entry<Attributes, AggregatorHandle<T>> entry : aggregatorLabels.entrySet()) {
+        boolean unmappedEntry = entry.getValue().tryUnmap();
+        if (unmappedEntry) {
+          // If able to unmap then remove the record from the current Map. This can race with the
+          // acquire but because we requested a specific value only one will succeed.
+          aggregatorLabels.remove(entry.getKey(), entry.getValue());
+        }
+        T accumulation = entry.getValue().accumulateThenReset(entry.getKey());
+        if (accumulation == null) {
+          continue;
+        }
+        instrumentProcessor.batch(entry.getKey(), accumulation);
+      }
+      return instrumentProcessor.completeCollectionCycle(epochNanos);
+    } finally {
+      collectLock.unlock();
+    }
+  }
+
+  @Override
+  public MetricDescriptor getMetricDescriptor() {
+    return metricDescriptor;
+  }
+}
