+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/internal/state/SynchronousMetricStorage.java
@@ -5,23 +5,14 @@
 
 package io.opentelemetry.sdk.metrics.internal.state;
 
-import io.opentelemetry.api.common.Attributes;
-import io.opentelemetry.context.Context;
 import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
 import io.opentelemetry.sdk.metrics.common.InstrumentDescriptor;
 import io.opentelemetry.sdk.metrics.data.MetricData;
 import io.opentelemetry.sdk.metrics.exemplar.ExemplarFilter;
 import io.opentelemetry.sdk.metrics.internal.aggregator.Aggregator;
-import io.opentelemetry.sdk.metrics.internal.aggregator.AggregatorHandle;
 import io.opentelemetry.sdk.metrics.internal.descriptor.MetricDescriptor;
-import io.opentelemetry.sdk.metrics.internal.view.AttributesProcessor;
 import io.opentelemetry.sdk.metrics.view.View;
 import io.opentelemetry.sdk.resources.Resource;
-import java.util.Map;
-import java.util.Objects;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.locks.ReentrantLock;
-import javax.annotation.Nullable;
 
 /**
  * Stores aggregated {@link MetricData} for synchronous instruments.
@@ -29,21 +20,14 @@ import javax.annotation.Nullable;
  * <p>This class is internal and is hence not for public use. Its APIs are unstable and can change
  * at any time.
  */
-public final class SynchronousMetricStorage<T> implements MetricStorage, WriteableMetricStorage {
-  private final MetricDescriptor metricDescriptor;
-  private final ConcurrentHashMap<Attributes, AggregatorHandle<T>> aggregatorLabels;
-  private final ReentrantLock collectLock;
-  private final Aggregator<T> aggregator;
-  private final InstrumentProcessor<T> instrumentProcessor;
-  private final AttributesProcessor attributesProcessor;
+public interface SynchronousMetricStorage extends MetricStorage, WriteableMetricStorage {
 
   /**
    * Constructs metric storage for a given synchronous instrument and view.
    *
    * @return The storage, or {@code null} if the instrument should not be recorded.
    */
-  @Nullable
-  public static <T> SynchronousMetricStorage<T> create(
+  public static <T> SynchronousMetricStorage create(
       View view,
       InstrumentDescriptor instrumentDescriptor,
       Resource resource,
@@ -62,132 +46,12 @@ public final class SynchronousMetricStorage<T> implements MetricStorage, Writeab
     // We won't be storing this metric.
     if (Aggregator.empty().equals(aggregator)) {
       // Short circuit with null-object.
-      return null;
+      return EmptyMetricStorage.INSTANCE;
     }
-    return new SynchronousMetricStorage<>(
+    return new DefaultSynchronousMetricStorage<>(
         metricDescriptor,
         aggregator,
         new InstrumentProcessor<>(aggregator, startEpochNanos),
         view.getAttributesProcessor());
-  }
-
-  SynchronousMetricStorage(
-      MetricDescriptor metricDescriptor,
-      Aggregator<T> aggregator,
-      InstrumentProcessor<T> instrumentProcessor,
-      AttributesProcessor attributesProcessor) {
-    this.metricDescriptor = metricDescriptor;
-    aggregatorLabels = new ConcurrentHashMap<>();
-    collectLock = new ReentrantLock();
-    this.aggregator = aggregator;
-    this.instrumentProcessor = instrumentProcessor;
-    this.attributesProcessor = attributesProcessor;
-  }
-
-  // This is a storage handle to use when the attributes processor requires
-  private final BoundStorageHandle lateBoundStorageHandle =
-      new BoundStorageHandle() {
-        @Override
-        public void release() {}
-
-        @Override
-        public void recordLong(long value, Attributes attributes, Context context) {
-          SynchronousMetricStorage.this.recordLong(value, attributes, context);
-        }
-
-        @Override
-        public void recordDouble(double value, Attributes attributes, Context context) {
-          SynchronousMetricStorage.this.recordDouble(value, attributes, context);
-        }
-      };
-
-  @Override
-  public BoundStorageHandle bind(Attributes attributes) {
-    Objects.requireNonNull(attributes, "attributes");
-    if (attributesProcessor.usesContext()) {
-      // We cannot pre-bind attributes because we need to pull attributes from context.
-      return lateBoundStorageHandle;
-    }
-    return doBind(attributesProcessor.process(attributes, Context.current()));
-  }
-
-  private BoundStorageHandle doBind(Attributes attributes) {
-    AggregatorHandle<T> aggregatorHandle = aggregatorLabels.get(attributes);
-    if (aggregatorHandle != null && aggregatorHandle.acquire()) {
-      // At this moment it is guaranteed that the Bound is in the map and will not be removed.
-      return aggregatorHandle;
-    }
-
-    // Missing entry or no longer mapped, try to add a new entry.
-    aggregatorHandle = aggregator.createHandle();
-    while (true) {
-      AggregatorHandle<?> boundAggregatorHandle =
-          aggregatorLabels.putIfAbsent(attributes, aggregatorHandle);
-      if (boundAggregatorHandle != null) {
-        if (boundAggregatorHandle.acquire()) {
-          // At this moment it is guaranteed that the Bound is in the map and will not be removed.
-          return boundAggregatorHandle;
-        }
-        // Try to remove the boundAggregator. This will race with the collect method, but only one
-        // will succeed.
-        aggregatorLabels.remove(attributes, boundAggregatorHandle);
-        continue;
-      }
-      return aggregatorHandle;
-    }
-  }
-
-  // Overridden to make sure attributes processor can pull baggage.
-  @Override
-  public void recordLong(long value, Attributes attributes, Context context) {
-    Objects.requireNonNull(attributes, "attributes");
-    attributes = attributesProcessor.process(attributes, context);
-    BoundStorageHandle handle = doBind(attributes);
-    try {
-      handle.recordLong(value, attributes, context);
-    } finally {
-      handle.release();
-    }
-  }
-
-  // Overridden to make sure attributes processor can pull baggage.
-  @Override
-  public void recordDouble(double value, Attributes attributes, Context context) {
-    Objects.requireNonNull(attributes, "attributes");
-    attributes = attributesProcessor.process(attributes, context);
-    BoundStorageHandle handle = doBind(attributes);
-    try {
-      handle.recordDouble(value, attributes, context);
-    } finally {
-      handle.release();
-    }
-  }
-
-  @Override
-  public MetricData collectAndReset(long startEpochNanos, long epochNanos) {
-    collectLock.lock();
-    try {
-      for (Map.Entry<Attributes, AggregatorHandle<T>> entry : aggregatorLabels.entrySet()) {
-        boolean unmappedEntry = entry.getValue().tryUnmap();
-        if (unmappedEntry) {
-          // If able to unmap then remove the record from the current Map. This can race with the
-          // acquire but because we requested a specific value only one will succeed.
-          aggregatorLabels.remove(entry.getKey(), entry.getValue());
-        }
-        T accumulation = entry.getValue().accumulateThenReset(entry.getKey());
-        if (accumulation == null) {
-          continue;
-        }
-        instrumentProcessor.batch(entry.getKey(), accumulation);
-      }
-      return instrumentProcessor.completeCollectionCycle(epochNanos);
-    } finally {
-      collectLock.unlock();
-    }
-  }
-
-  @Override
-  public MetricDescriptor getMetricDescriptor() {
-    return metricDescriptor;
   }
 }
